Break long variant filenames into path chunks
If a variant has a large set of options associated with it, the generated
filename will be too long, causing Errno::ENAMETOOLONG to be raised. This
change adds intermediary paths that retain the uniqueness of the variant
while not raising the aforementioned error until the overall 1024 path
limit is exceeded. Fixes #30662.